tools/magnus_talent_cli/requirements.txt
openai>=1.12.0
supabase-py>=2.5.0
python-dotenv>=1.0.1
rich>=13.7.0
requests>=2.32.3
